+++ core/performance/liftingPerformance.ts
@@ -1,35 +1,17 @@
-// NOT DONE YET!!!!
+// TODO: NOT DONE YET!!!!
 // zaka no seas vago y trabaja
 
 /*
 CALCULATE RUNNING / WALKING PERFORMANCE
 */
 
-import { default as OneRepetitionMax } from "@/core/physicalHealth/oneRepMax"
+import { default as OneRepetitionMax } from "@/core/physicalHealth/OneRepMax"
+import CreateComponentDataUtilities from "@/core/tools/OpenHealthDataBuilder";
 
-// LAST UPDATE TO THIS FUNCTION, ITS DATA, ITS CALCULATIONS, OR ANYTHING THAT DOES AFFECT THE RESULT
-// Changes that do not affect the result, like just bug-fixes, performance improvments, code-legibility improvments, or that kind of stuff, do not need to bump the date.
-const UPDATED: string = "01/07/2024";
-// ANY SOURCE THAT HAS BEEN USED TO DEVELOP THE CALCULATIONS / DATA PROVIDED or that BACKS IT UP.
-const SOURCE: string = ""; // NEED TO PROVIDE SOURCE
-
-/**
- * Get all the sources of information used to develop the function, it's data, contents, returns, and etc...
- * @returns A single string with all the URLs, separated by "and" in case there's more than one - e.g. "https://coolsite.com and https://example.source"
-*/
-
-export function getSource() {
-    return SOURCE;
-}
-
-/**
- * Get the date of the last update made to this function (considering "update" any change to it's sources, calculations, data, results, etc... but not trivial, performance, code cleaning or similar changes.)
- * @returns A string with the date of the last update, using the DD/MM/YYYY format.
-*/
-
-export function getLastUpdate() {
-    return UPDATED;
-}
+export const { getSource, getLastUpdate } = CreateComponentDataUtilities(
+    "01/07/2024",
+    "" // FIXME: NEED TO PROVIDE SOURCE
+)
 
 interface LIFTING_Response {
     result: number;