+++ core/physicalHealth/MetabolicEquivalentOfTask.ts
@@ -2,41 +2,13 @@
 CALCULATE THE METABOLIC EQUIVALENT OF A TASK
 */
 
-// LAST UPDATE TO THIS FUNCTION, ITS DATA, ITS CALCULATIONS, OR ANYTHING THAT DOES AFFECT THE RESULT
-// Changes that do not affect the result, like just bug-fixes, performance improvments, code-legibility improvments, or that kind of stuff, do not need to bump the date.
-const UPDATED: string = "01/07/2024";
-// ANY SOURCE THAT HAS BEEN USED TO DEVELOP THE CALCULATIONS / DATA PROVIDED or that BACKS IT UP.
-const SOURCE: string = "https://en.wikipedia.org/wiki/Metabolic_equivalent_of_task > https://www.ahajournals.org/doi/10.1161/CIRCULATIONAHA.107.185649 and https://www.inchcalculator.com/calories-burned-weight-lifting-calculator/";
+import CreateComponentDataUtilities from "@/core/tools/OpenHealthDataBuilder";
+import { OpenHealthResponseVersatile } from "@/core/types/OpenHealthResponse";
 
-
-/**
- * Get all the sources of information used to develop the function, it's data, contents, returns, and etc...
- * @returns A single string with all the URLs, separated by "and" in case there's more than one - e.g. "https://coolsite.com and https://example.source"
-*/
-
-export function getSource() {
-    return SOURCE;
-}
-
-/**
- * Get the date of the last update made to this function (considering "update" any change to it's sources, calculations, data, results, etc... but not trivial, performance, code cleaning or similar changes.)
- * @returns A string with the date of the last update, using the DD/MM/YYYY format.
-*/
-
-export function getLastUpdate() {
-    return UPDATED;
-}
-
-interface MET_Response {
-    result: number;
-    subject?: {
-        age: number;
-        gender: "male" | "female";
-        intensity: "superlow" | "very_low" | "low" | "low_to_mid" | "mid" | "mid_to_high" | "not_too_high" | "high" | "higher" | "very_high" | "very_high_to_intense" | "not_too_intense" | "a_bit_intense" | "intense" | "pretty_intense" | "very_intense" | "really_intense";
-    };
-    context?: string;
-    explanation?: string;
-}
+export const { getSource, getLastUpdate } = CreateComponentDataUtilities(
+    "01/07/2024",
+    "https://en.wikipedia.org/wiki/Metabolic_equivalent_of_task > https://www.ahajournals.org/doi/10.1161/CIRCULATIONAHA.107.185649 and https://www.inchcalculator.com/calories-burned-weight-lifting-calculator/"
+)
 
 /**
  * Returns the Metabolic Equivalent of a Task
@@ -45,7 +17,7 @@ interface MET_Response {
  * @param intensity The intensity of the activity. It's a very long array of options, see the documentation for more info on which one to choose.
  * @param provideContext Whether to provide a brief contextualization about the result.
  * @param provideExplanation Whether to provide a detailed explanation about what the calculation means.
- * @returns A number with the Metabolic Equivalent of the Task if neither provideContext nor provideExplanation are true, otherwise returns a MET_Response object.
+ * @returns A number with the Metabolic Equivalent of the Task if neither provideContext nor provideExplanation are true, otherwise returns a `OpenHealthResponseVersatile` object.
 */
 
 export default function calculateMetabolicEquivalentOfTask(
@@ -54,7 +26,7 @@ export default function calculateMetabolicEquivalentOfTask(
     intensity: "superlow" | "very_low" | "low" | "low_to_mid" | "mid" | "mid_to_high" | "not_too_high" | "high" | "higher" | "very_high" | "very_high_to_intense" | "not_too_intense" | "a_bit_intense" | "intense" | "pretty_intense" | "very_intense" | "really_intense",
     provideContext?: boolean,
     provideExplanation?: boolean
-): MET_Response {
+): OpenHealthResponseVersatile {
     let mets: number; // MET - Metabolic Equivalent of Task
 
     // Assign METs based on intensity level
@@ -96,7 +68,7 @@ export default function calculateMetabolicEquivalentOfTask(
         throw new Error("CALCULATION ERROR! Intensinty is not valid")
     }
 
-    const response: MET_Response = {
+    const response: OpenHealthResponseVersatile = {
         result: mets,
     };