+++ core/physicalHealth/bodymassindex.ts
@@ -3,43 +3,12 @@ CALCULATE BODY MASS INDEX
 */
 
 import { OpenHealthResponse } from "@/core/types/OpenHealthResponse";
+import CreateComponentDataUtilities from "@/core/tools/OpenHealthDataBuilder";
 
-// LAST UPDATE TO THIS FUNCTION, ITS DATA, ITS CALCULATIONS, OR ANYTHING THAT DOES AFFECT THE RESULT
-// Changes that do not affect the result, like just bug-fixes, performance improvments, code-legibility improvments, or that kind of stuff, do not need to bump the date.
-const UPDATED: string = "10/08/2024";
-// ANY SOURCE THAT HAS BEEN USED TO DEVELOP THE CALCULATIONS / DATA PROVIDED or that BACKS IT UP.
-const SOURCE: string = "https://www.cdc.gov/growthcharts/data/set1clinical/cj41c024.pdf and https://www.cdc.gov/growthcharts/data/set1clinical/cj41c023.pdf";
-// i initially created this with https://www.cdc.gov/healthyweight/assessing/bmi/adult_bmi/index.html#Interpreted, but the link doesnt work now :v
-
-/**
- * Get all the sources of information used to develop the function, it's data, contents, returns, and etc...
- * @returns A single string with all the URLs, separated by "and" in case there's more than one - e.g. "https://coolsite.com and https://example.source"
-*/
-
-export function getSource() {
-    return SOURCE;
-}
-
-/**
- * Get the date of the last update made to this function (considering "update" any change to it's sources, calculations, data, results, etc... but not trivial, performance, code cleaning or similar changes.)
- * @returns A string with the date of the last update, using the DD/MM/YYYY format.
-*/
-
-export function getLastUpdate() {
-    return UPDATED;
-}
-
-/* interface BMIResponse {
-    result: number;
-    subject?: {
-        age: number;
-        gender: "male" | "female";
-        weight: number;
-        height: number;
-    };
-    context?: string;
-    explanation?: string;
-} */
+export const { getSource, getLastUpdate } = CreateComponentDataUtilities(
+    "10/08/2024",
+    "https://www.cdc.gov/growthcharts/data/set1clinical/cj41c024.pdf and https://www.cdc.gov/growthcharts/data/set1clinical/cj41c023.pdf"
+)
 
 /**
  * Represents percentiles data for Body Mass Index (BMI) for individuals between 2 and 20 years of age.