+++ core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.3
 
 - (WIP) Replace having an interface for each function with a generic `OpenHealthResponse`.
+- Replace duplicate code with a "constructor" function for `getSource` and `getLastUpdate`.
 - Update `physicalHealth` / `Body Mass Index`, added support for under 20 years of age calculations.
 - Some work around with DOCS and READMEs.
 - Replace `./` routing with `@/` for imports.